Make `cagkftab.py` pytest-ready
